digit_recognition/requirements.txt
Pillow==7.2.0
tensorflow==2.3.1
Keras==2.4.3